[common] fix ci build error

+++ source/ngf-common/block-alloc.c
@@ -112,7 +112,7 @@ ngfi_block_allocator* ngfi_blkalloc_create(uint32_t requested_block_size, uint32
 
   const uint32_t unaligned_block_size = requested_block_size + sizeof(ngfi_blkalloc_block);
   const uint32_t aligned_block_size =
-      unaligned_block_size + (NGFI_MAX_ALIGNMENT - (unaligned_block_size & (NGFI_MAX_ALIGNMENT - 1u)));
+      unaligned_block_size + (uint32_t)(NGFI_MAX_ALIGNMENT - (unaligned_block_size & (NGFI_MAX_ALIGNMENT - 1u)));
 
   allocator->block_size = aligned_block_size;
   allocator->block_size_user = requested_block_size;
